Dispersion Types

RetrievalToolbox.SimplePolynomialDispersion — Type

spectral_window::AbstractSpectralWindow: Spectral window

coefficients::Vector{T2} where T2<:AbstractFloat: Polynomial coefficients, ordered low to high

index::Vector{Int64}: Spectral samples within the current spectral window!

ww::Vector{T2} where T2<:AbstractFloat: Wavelengths corresponding to the index

detector_samples::AbstractRange: The full detector grid corresponding to the L1B spectral samples

ww_unit::Union{Unitful.Units{U, 𝐋} where U, Unitful.Units{U, 𝐋^-1} where U}: Unitful unit of all wavelength and coefficient quantities

A type to represent a polynomial dispersion, where each spectral sample is related to wavelength or wavenumber through a polynomial of arbitrary degree.

RetrievalToolbox.SimplePolynomialDispersion — Method

Generates a SimplePolynomialDispersion object from a vector of polynomial coefficients (coeffs), some detector samples given as an AbstractRange object like 1:1016, (detector_samples), and a spectral window of type AbstractSpectralWindow that this dispersion object will refer to.

SimplePolynomialDispersion(
    coeffs::AbstractVector,
    detector_samples::AbstractRange,
    spectral_window::AbstractSpectralWindow
) -> SimplePolynomialDispersion